🚚 rename `inversion` module to `core`
