+++ cherab/inversion/__init__.py
@@ -1,8 +1,8 @@
 """Subpackage for Inversion Problem."""
+from .core import _SVDBase, compute_svd
 from .gcv import GCV
-from .inversion import _SVDBase, compute_svd
 from .lcurve import Lcurve
 from .mfr import Mfr
 
-__all__ = ["_SVDBase", "compute_svd", "Lcurve", "GCV", "Mfr"]
+__all__ = ["compute_svd", "_SVDBase", "Lcurve", "GCV", "Mfr"]
 __version__ = "0.1.dev0"
