+++ cherab/inversion/gcv.py
@@ -7,7 +7,7 @@ from matplotlib.axes import Axes
 from matplotlib.figure import Figure
 from numpy.typing import NDArray
 
-from .inversion import _SVDBase
+from .core import _SVDBase
 
 __all__ = ["GCV"]
